src/ol_openedx_canvas_integration/pyproject.toml
[project]
name = "ol-openedx-canvas-integration"
version = "0.5.0"
description = "An Open edX plugin to add canvas integration support"
authors = [
  {name = "MIT Office of Digital Learning"}
]
license = "BSD-3-Clause"
readme = "README.rst"
requires-python = ">=3.11"
dependencies = [
  "Django>2.0",
  "celery>=4.4.7",
  "djangorestframework>=3.14.0",
  "edx-django-utils>4.0.0",
  "edx-drf-extensions>=10.0.0",
  "edx-opaque-keys",
  "openedx-events",
]

[project.entry-points."lms.djangoapp"]
ol_openedx_canvas_integration = "ol_openedx_canvas_integration.app:CanvasIntegrationConfig"

[project.entry-points."cms.djangoapp"]
ol_openedx_canvas_integration = "ol_openedx_canvas_integration.app:CanvasIntegrationConfig"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
  "ol_openedx_canvas_integration/**/*.py",
  "ol_openedx_canvas_integration/static/**/*",
  "ol_openedx_canvas_integration/templates/**/*",
]
exclude = [
  "ol_openedx_canvas_integration/**/test_*"
]


[tool.hatch.build.targets.sdist]
include = [
  "ol_openedx_canvas_integration/**/*",
  "README.rst",
  "pyproject.toml",
]
exclude = [
  "ol_openedx_canvas_integration/**/test_*"
]
